requirements.txt
python-dotenv
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
streamlit
langchain_community